CRUD for ApiUser

+++ app/handlers/device_handler.py
@@ -1,4 +1,3 @@
-# handlers/device_handler.py
 import logging
 from aiohttp import web
 from services import DeviceService
@@ -11,8 +10,9 @@ class DeviceHandler:
 
     async def create_device(self, request):
         try:
+            user = request.user
             data = await request.json()
-            print(data)
+            data['api_user_id'] = user.id
             device = await self.device_service.create_device(**data)
             return web.json_response({'id': device.id})
         except Exception as e:
@@ -21,20 +21,26 @@ class DeviceHandler:
 
     async def get_device(self, request):
         try:
+            user = request.user
             device_id = int(request.match_info['id'])
-            device = self.device_service.get_device(device_id)
-            if device:
+            device = await self.device_service.get_device(device_id)
+            if device and device.api_user.id == user.id:
                 return web.json_response(device)
-            return web.Response(status=404, text="Device not found")
+            return web.Response(status=404, text="Device not found or not owned by user")
         except Exception as e:
             self.logger.error(f"Error getting device: {e}")
             return web.Response(status=500, text=str(e))
 
     async def update_device(self, request):
         try:
+            user = request.user
             device_id = int(request.match_info['id'])
+            device = await self.device_service.get_device(device_id)
+            if not device or device.api_user.id != user.id:
+                return web.Response(status=404, text="Device not found or not owned by user")
+            
             data = await request.json()
-            self.device_service.update_device(device_id, data)
+            await self.device_service.update_device(device_id, data)
             return web.Response(status=200)
         except Exception as e:
             self.logger.error(f"Error updating device: {e}")
@@ -42,8 +48,13 @@ class DeviceHandler:
 
     async def delete_device(self, request):
         try:
+            user = request.user
             device_id = int(request.match_info['id'])
-            self.device_service.delete_device(device_id)
+            device = await self.device_service.get_device(device_id)
+            if not device or device.api_user.id != user.id:
+                return web.Response(status=404, text="Device not found or not owned by user")
+            
+            await self.device_service.delete_device(device_id)
             return web.Response(status=200)
         except Exception as e:
             self.logger.error(f"Error deleting device: {e}")
